feat: moved sync settings for summarization and screenshots to urls
