+++ includes/cron/class-urlslab-cron-screenshots.php
@@ -17,25 +17,22 @@ class Urlslab_Cron_Screenshots extends Urlslab_Cron {
 		if ( ! Urlslab_User_Widget::get_instance()->is_widget_activated( Urlslab_Widget_Urls::SLUG ) || ! $this->init_client() ) {
 			return false;
 		}
-		$widget = Urlslab_User_Widget::get_instance()->get_widget( Urlslab_Widget_General::SLUG );
+		$widget = Urlslab_User_Widget::get_instance()->get_widget( Urlslab_Widget_Urls::SLUG );
 		global $wpdb;
 
-		$query_data = array();
+		$query_data            = array();
 		$sql_where_http_status = '';
-		$use_index = '';
-		if (
-			Urlslab_User_Widget::get_instance()->is_widget_activated( Urlslab_Widget_Urls::SLUG )
-			&& Urlslab_User_Widget::get_instance()->get_widget( Urlslab_Widget_Urls::SLUG )->get_option( Urlslab_Widget_Urls::SETTING_NAME_VALIDATE_LINKS )
-		) {
+		$use_index             = '';
+		if ( $widget->get_option( Urlslab_Widget_Urls::SETTING_NAME_VALIDATE_LINKS ) ) {
 			$query_data[]          = Urlslab_Data_Url::HTTP_STATUS_OK;
 			$sql_where_http_status = ' http_status = %d AND';
-			$use_index = ' USE INDEX (idx_scr_cron)';
+			$use_index             = ' USE INDEX (idx_scr_cron)';
 		}
 
 		$query_data[]    = Urlslab_Data_Url::SCR_STATUS_NEW;
 		$where_status_or = '';
 
-		switch ( $widget->get_option( Urlslab_Widget_General::SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL ) ) {
+		switch ( $widget->get_option( Urlslab_Widget_Urls::SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL ) ) {
 			case DomainDataRetrievalDataRequest::RENEW_FREQUENCY_ONE_TIME:
 				break;
 			case DomainDataRetrievalDataRequest::RENEW_FREQUENCY_DAILY:
@@ -80,6 +77,7 @@ class Urlslab_Cron_Screenshots extends Urlslab_Cron {
 		);
 		if ( empty( $url_rows ) ) {
 			$this->lock( 300, Urlslab_Cron::LOCK );
+
 			return false;
 		}
 
@@ -110,7 +108,7 @@ class Urlslab_Cron_Screenshots extends Urlslab_Cron {
 
 		$request = new DomainDataRetrievalDataRequest();
 		$request->setUrls( $url_names );
-		$request->setRenewFrequency( $widget->get_option( Urlslab_Widget_General::SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL ) );
+		$request->setRenewFrequency( $widget->get_option( Urlslab_Widget_Urls::SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL ) );
 
 		try {
 			$urlslab_screenshots = $this->client->getSnapshots( $request );
