+++ includes/widget/class-urlslab-widget-general.php
@@ -5,13 +5,11 @@
 use Urlslab_Vendor\OpenAPI\Client\Configuration;
 use Urlslab_Vendor\OpenAPI\Client\Urlslab\ApikeyApi;
 use Urlslab_Vendor\GuzzleHttp;
-use Urlslab_Vendor\OpenAPI\Client\Model\DomainDataRetrievalDataRequest;
 
 class Urlslab_Widget_General extends Urlslab_Widget {
 	public const SLUG = 'general';
 
 	public const SETTING_NAME_URLSLAB_API_KEY = 'urlslab-api-key';
-	public const SETTING_NAME_SUMMARIZATION_REFRESH_INTERVAL = 'urlslab-refresh-sum';
 	public const SETTING_NAME_URLSLAB_CREDITS = 'urlslab-credits';
 	const SETTING_NAME_DOMAIN_BLACKLIST = 'urlslab-url-blacklist';
 	const SETTING_NAME_CLASSNAMES_BLACKLIST = 'urlslab-classnames-blacklist';
@@ -23,12 +21,6 @@ class Urlslab_Widget_General extends Urlslab_Widget {
 	const SETTING_NAME_USE_HTACCESS = 'urlslab-cache-htaccess';
 	const SETTING_NAME_HTACCESS_VERSION = 'urlslab-htaccess-version';
 
-	const SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL = 'urlslab-scr-refresh';
-	const SETTING_NAME_SHEDULE_SCRRENSHOT = 'urlslab-scr-schedule';
-	public const SCHEDULE_SHORTCODE = 'S';
-	public const SCHEDULE_ALL_INTERNALS = 'I';
-	public const SCHEDULE_ALL = 'A';
-	public const SCHEDULE_NEVER = 'N';
 	const SETTING_NAME_IP_ANONYMIZATION = 'urlslab-ip-anonym';
 	const SETTING_NAME_IGNORE_PARAMETERS = 'urlslab-param-blacklist';
 	const SETTING_NAME_DOMAIN_WHITELIST = 'urlslab-domain-whitelist';
@@ -163,105 +155,6 @@ class Urlslab_Widget_General extends Urlslab_Widget {
 			false,
 			null,
 			'api'
-		);
-
-		$this->add_options_form_section(
-			'summarization',
-			function () {
-				return __( 'Summarizations', 'urlslab' );
-			},
-			function () {
-				return __( 'Summarizations are generated by the URLsLab service. Adjust the synchronization frequency between URLsLab service and the plugin here.', 'urlslab' );
-			},
-			array( self::LABEL_PAID )
-		);
-		$this->add_option_definition(
-			self::SETTING_NAME_SUMMARIZATION_REFRESH_INTERVAL,
-			2419200,
-			false,
-			function () {
-				return __( 'Synchronization with URLsLab Service', 'urlslab' );
-			},
-			function () {
-				return __( 'The frequency of summary sync with URLsLab differs from the rate at which URLsLab creates page summaries. Even with background data sync via cron, significant computation time may be used.', 'urlslab' );
-			},
-			self::OPTION_TYPE_LISTBOX,
-			function () {
-				return array(
-					604800           => __( 'Weekly', 'urlslab' ),
-					2419200          => __( 'Monthly', 'urlslab' ),
-					7257600          => __( 'Quarterly', 'urlslab' ),
-					31536000         => __( 'Yearly', 'urlslab' ),
-					self::FREQ_NEVER => __( 'Never', 'urlslab' ),
-				);
-			},
-			function ( $value ) {
-				return is_numeric( $value ) && 0 < $value;
-			},
-			'summarization',
-		);
-
-		$this->add_options_form_section(
-			'screenshots',
-			function () {
-				return __( 'Screenshots', 'urlslab' );
-			},
-			function () {
-				return __( 'The URLsLab Screenshot service, is a premium feature of the plugin - generates screenshot of any URL.', 'urlslab' );
-			},
-			array( self::LABEL_PAID )
-		);
-		$this->add_option_definition(
-			self::SETTING_NAME_SHEDULE_SCRRENSHOT,
-			self::SCHEDULE_SHORTCODE,
-			false,
-			function () {
-				return __( 'URL types', 'urlslab' );
-			},
-			function () {
-				return __( 'Select the types of URLs for which should be generated screenshot.', 'urlslab' );
-			},
-			self::OPTION_TYPE_LISTBOX,
-			function () {
-				return array(
-					self::SCHEDULE_NEVER         => __( 'None', 'urlslab' ),
-					self::SCHEDULE_SHORTCODE     => __( 'When URL (external or internal) is used in screenshot shortcode', 'urlslab' ),
-					self::SCHEDULE_ALL_INTERNALS => __( 'Every internal URL', 'urlslab' ),
-					self::SCHEDULE_ALL           => __( 'All URLs in database', 'urlslab' ),
-				);
-			},
-			function ( $value ) {
-				return is_string( $value );
-			},
-			'screenshots',
-		);
-		$this->add_option_definition(
-			self::SETTING_NAME_SCREENSHOT_REFRESH_INTERVAL,
-			DomainDataRetrievalDataRequest::RENEW_FREQUENCY_MONTHLY,
-			false,
-			function () {
-				return __( 'Screenshot Synchronization', 'urlslab' );
-			},
-			function () {
-				return __( 'Choose the frequency at which plugin synchronize screenshots with URLsLab service.', 'urlslab' );
-			},
-			self::OPTION_TYPE_LISTBOX,
-			function () {
-				return array(
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_NO_SCHEDULE => __( 'Never', 'urlslab' ),
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_ONE_TIME    => __( 'Only once', 'urlslab' ),
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_DAILY       => __( 'Daily', 'urlslab' ),
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_WEEKLY      => __( 'Weekly', 'urlslab' ),
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_MONTHLY     => __( 'Monthly', 'urlslab' ),
-					DomainDataRetrievalDataRequest::RENEW_FREQUENCY_YEARLY      => __( 'Yearly', 'urlslab' ),
-				);
-			},
-			function ( $value ) {
-				$obj = new DomainDataRetrievalDataRequest();
-
-				return in_array( $value, $obj->getRenewFrequencyAllowableValues() );
-			},
-			'screenshots',
 		);
